+++ projects/angular2-smart-table/src/lib/components/cell/cell-edit-mode/edit-cell-default.ts
@@ -11,6 +11,7 @@ export class EditCellDefault {
   @Input() inputClass: string = '';
 
   @Output() edited = new EventEmitter<void>();
+  @Output() stopEditing = new EventEmitter<void>();
 
   onEdited(): boolean {
     this.edited.emit();
@@ -18,7 +19,7 @@ export class EditCellDefault {
   }
 
   onStopEditing(): boolean {
-    this.cell.getRow().isInEditing = false;
+    this.stopEditing.emit();
     return false;
   }